Fix the inferred typing for useAuthMutation
Before, we were casting TVariables as TData initially. This caused the mutation function to have a required first argument "variables" that matched the return type of the mutation function provided (when no explicit argument types are provided, i.e. estimateShipping, calculateOrder, etc). By changing the TVariables to match the initial type of UseMutationOptions - it should be cast as "void", unless of course there is an explicit variable required for the mutation function, which should be inherited by the type arguments of the mutation function provided.